Remove SHA256 in favor of an extension on Data
